+++ src/packer/core.py
@@ -504,6 +504,13 @@ class Packer():
             if not self.pypi_api_token:
                 building_cmd.insert(3, '--wheel')
             building_done = self._Popen(building_cmd, waiting_for_building)
+
+            for pattern in exclusions:
+                if Path('build').match(pattern):
+                    break
+            else:
+                self.print_and_log('Removing build directory...')
+                rmtree('build')
             
             if self.run_pyinstaller:
                 self.print_and_log('Bundling the program using PyInstaller...')